Moved files around for builtin threads
Eventually cactus_rt will spawn a few threads by default. The definition
of these threads should be separated from the definition of the base
framework API.
