+++ include/cactus_rt/base/latency_tracker.h
@@ -1,5 +1,5 @@
-#ifndef CACTUS_RT_LATENCY_TRACKING_H_
-#define CACTUS_RT_LATENCY_TRACKING_H_
+#ifndef CACTUS_RT_BASE_LATENCY_TRACKING_H_
+#define CACTUS_RT_BASE_LATENCY_TRACKING_H_
 
 #include <array>
 #include <cstdint>
