+++ include/cactus_rt/base/utils.h
@@ -1,5 +1,5 @@
-#ifndef CACTUS_RT_UTILS_H_
-#define CACTUS_RT_UTILS_H_
+#ifndef CACTUS_RT_BASE_UTILS_H_
+#define CACTUS_RT_BASE_UTILS_H_
 
 #include <cstdint>
 #include <ctime>
