+++ src/base/signal_handler.cc
@@ -1,4 +1,4 @@
-#include "cactus_rt/signal_handler.h"
+#include "cactus_rt/base/signal_handler.h"
 
 namespace cactus_rt {
 sem_t signal_semaphore;
